# VersionManifestAddendum caching, runtime log debug flag