--- a/ModTek/Patches/SimGameConstants.cs
+++ b/ModTek/Patches/SimGameConstants.cs
@@ -9,7 +9,8 @@ namespace ModTek
     [HarmonyPatch(MethodType.Normal)]
     public static class SimGameConstants_FromJSON
     {
-        public static bool Prepare() { return ModTek.Enabled; }
+        public static bool Prepare() => ModTek.Enabled && ModTek.Config.EnableDebugLogging;
+
         public static void Postfix(SimGameConstants __instance, string json)
         {
             RLog.M.TWL(0, "SimGameConstants.FromJSON");