--- a/ModTek/Patches/UIManagerPatches.cs
+++ b/ModTek/Patches/UIManagerPatches.cs
@@ -27,17 +27,28 @@ namespace ModTek.Patches
         {
             DataManager dm = SceneSingletonBehavior<DataManagerUnityInstance>.Instance.DataManager;
 
+
             Stopwatch sw = new Stopwatch();
             sw.Start();
-            RLog.LogWrite("Forcing refresh of BTRL Manifest...\n");
+            if (ModTek.Config.EnableDebugLogging)
+                RLog.LogWrite("Forcing refresh of BTRL Manifest...\n");
+
             Traverse.Create(dm.ResourceLocator).Property("manifest").SetValue(VersionManifestUtilities.LoadDefaultManifest());
-            RLog.LogWrite($"Refresh of BTRL manifest complete in {sw.ElapsedMilliSeconds}ms \n");
+
+            sw.Stop();
+            if (ModTek.Config.EnableDebugLogging)
+                RLog.LogWrite($"Refresh of BTRL manifest complete in {sw.ElapsedMilliSeconds}ms \n");
 
             sw.Restart();
+            if (ModTek.Config.EnableDebugLogging)
+                RLog.LogWrite("Forcing refresh of TypedEntities to prevent Shadowhawk DLC bug...\n");
+
             Traverse refreshTypedEntriesT = Traverse.Create(dm.ResourceLocator).Method("RefreshTypedEntries");
-            RLog.LogWrite("Forcing refresh of TypedEntities to prevent Shadowhawk DLC bug...\n");
             refreshTypedEntriesT.GetValue();
-            RLog.LogWrite($"Forced refresh of TypedEntities complete in {sw.ElapsedMilliSeconds}ms\n");
+
+            sw.Stop();
+            if (ModTek.Config.EnableDebugLogging)
+                RLog.LogWrite($"Forced refresh of TypedEntities complete in {sw.ElapsedMilliSeconds}ms\n");
         }
     }
 }