This PR covers two minor performance increases. When KMission added his runtime logger, he used it to log all manifest entries that were added. This adds several seconds to my modtek load, so I've put this behind a configuration gate 'EnableDebugLogging' in config.json. To accommodate this, I've repointed the configuration file to be found at ModTek/ instead of .modtek/, as /.modtek is created at runtime.

The bigger change is that this mod implements a dictionary cache for VersionManifestAddendum.GetAddendumContainingEntry calls. The vanilla code for this uses LINQ to do Contains() calls, which due to the nested nature of VersionManifestAddendum -> VersionManifestEntry is fairly painful. Instead this patch creates a cache of the VME values when the addendum is first queried, and then uses that cache for lookup. On my modtek this shaved ~4-8s off, as the VMA.GACE calls went from ~200ms -> ~75ms or ~600ms -> 100ms. These calls are performed repeatedly during the load.